Rewritten the function responsible for heroes tunnelling. Also rewritten a digging routine, and renamed some variables.

+++ trunk/keeperfx/src/dungeon_data.c
@@ -35,12 +35,12 @@ struct Dungeon *get_players_num_dungeon_ptr(long plyr_idx,const char *func_name)
     plyr_num = player->id_number;
     if (player_invalid(player) || (plyr_num < 0) || (plyr_num >= DUNGEONS_COUNT))
     {
-        ERRORMSG("%s: Tried to get players %ld nonexisting dungeon %ld!",func_name,(long)plyr_idx,(long)plyr_num);
+        ERRORMSG("%s: Tried to get players %d nonexisting dungeon %d!",func_name,(int)plyr_idx,(int)plyr_num);
         return INVALID_DUNGEON;
     }
     if (plyr_num != player->id_number)
     {
-        WARNDBG(7,"%s: Player number(%ld) differ from index(%ld)!",func_name,(long)plyr_num,plyr_idx);
+        WARNDBG(7,"%s: Player number(%d) differ from index(%d)!",func_name,(int)plyr_num,(int)plyr_idx);
     }
     return &(game.dungeon[plyr_num]);
 }
